+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Next]
 
 ### Added
-- ISpanFormattable implementation
+- ISpanFormattable and IUtf8SpanFormattable implementation
 
 ### Changed
 - Rename SafeDivide to Split and move in NodaMoney namespace